Update around database func.
