+++ Aoi/profiles.py
@@ -1,24 +1,28 @@
-from discord.ext import commands
 from discord import app_commands
-import discord
-import asyncio
-from emoji import demojize
 from discord.app_commands import locale_str as _T
-from . import (get_database_url,
-               update_profile_id,
-               update_prefix,
-               update_emoji_id,
-               update_freshman_id,
-               convert_user_to_mention,
-               convert_channel_to_mention,
-               convert_role_to_mention,
-               get_pre_pro_log_fre_sen_emo_ten_lim_adj,
-               get_pro_log_fre_sen_emo,
-               update_log_id,
-               update_senior_id,
-               get_profile_id)
+from discord.ext import commands
 from discord.utils import get
+from emoji import demojize
+import asyncio
+import discord
 
+from . import (
+    convert_channel_to_mention,
+    convert_role_to_mention,
+    convert_user_to_mention,
+    get_database_url,
+)
+from .database import (
+    get_pre_pro_log_fre_sen_emo_ten_lim_adj,
+    get_pro_log_fre_sen_emo,
+    get_profile_id,
+    update_emoji_id,
+    update_freshman_id,
+    update_log_id,
+    update_prefix,
+    update_profile_id,
+    update_senior_id,
+)
 
 DATABASE_URL = get_database_url()
