+++ pyproject.toml
@@ -16,6 +16,7 @@ selenium = "^4.4.3"
 Pillow = "^9.2.0"
 requests-oauthlib = "^1.3.1"
 emoji = "^2.0.0"
+bottle = "^0.12.23"
 
 [tool.poetry.dev-dependencies]
 ipykernel = "^6.15.1"
